requirements.txt
beautifulsoup4==4.9.3
Django==3.1.2
django-environ
psycopg2>=2.7,<3.0
psycopg2-binary
pandas==1.1.3
requests==2.24.0
pykakasi
mecab-python3
python-Levenshtein
ipython
geopy